Py: Create map_template.py
